Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

> Also, I do not think it is a good practice to actually perform business logic inside of the constructor (i.e. send a ping). I haven't thought it completely through
> yet, but I do feel like there is potentially some threading concerns there due to visibility.

I hear what you say about doing too much in the constructor of AsyncPingEntry. I can definitely move that code out of there to make it more explicit.

> To keep concurrency simpler I would suggest just using synchronize blocks instead of trying to use the concurrent map, especially if you keep this idea 
> of "polling”
> ...
> I see what you are trying to do with the "continuous polling concept", but I think that might be over complicating the example. The idea here (I think) is to
> illustrate what the recommend approach is for handling long-running operations started by a REST call - correct? If so, then I would remove the timer logic as > that adds complexity. Perhaps change your methods to be "startPing", and then have a "getPingResult" method instead.

The changes in patch set 2 of gerrit 6992 allows for a continuous ping in the background, so the timer thread is not only polling for results, but also
initiating a new echo for the entries that got an answer from rpc (positive or negative). I am not keen on the complexity it brought in either. I can
remove the timer thread and use ‘synchronized’ instead of concurrentMap for sure. Still, I think I will need something to remove entries from the map
otherwise I would not know when to do it. So, do you you think its okay to keep it as 'create'/'get’/‘remove'?

Patch-set: 2