Update patch set 2
Patch Set 2:

I think a startPing, getPingResult, and clearPingResult is fine. Perhaps add a comment that says something like "//In a real scenario, you would want to consider a cache which automatically drops entries" etc.

The only other option is to use some of googles caches or a weakhash map etc... but I think that will only add complexity in this case.

Patch-set: 2