Update patch set 4
Patch Set 4:

Hi Prasanth,
the main difference from your approach is that WAB application developers do not need to decide which packages with filters to import. This would simplify switching from current CorsFilter implementation if there would be a need to do so, even if this change was required only for certain distribution.

Patch-set: 4