Update patch set 4
Patch Set 4:

Moving the filter from catalina bundle fragment to a standalone module requires that we make the class visible. If the import is a concern, you can leave the filter class as is and move the relevant filter related registrations to a web-fragment.xml

What you are attempting to put in looks a lot similar to web-fragment feature of Servlets. Can you clarify the other questions I had in the previous comment.

Patch-set: 4