BUG-509: use concurrent hash-trie Map
This patch switches to using concurrent hash-trie map for metadata
tracking. It has the nice feature of having O(1) concurrent snapshots,
which should give us better performance when dealing with large children
bases.

Now this is a trade-off, so we add some logic to manage this -- which is
where MapAdaptor comes into play. It sees each ingress/egress of the
nodes and changes the underlying implementation accordingly. The cutoff
number of entries is controller via the
org.opendaylight.yangtools.datatree.triemap.cutoff property, which has
to be a positive integer. Default is 100 entries.

Change-Id: I2ddb81a1296cc17528a605eeaeda1f303560fb55
Signed-off-by: Robert Varga <rovarga@cisco.com>