Update patch set 2
Patch Set 2:

Please comment on the status of this gerrit - it looks like this work was completed at one point, but never merged in. 

We should merge it in (unfortunately there were some major changes in the RestConfImpl class, which is no longer a .xtend file). If you need help rebasing / merging with new changes there please reachout.

Patch-set: 2
Label: Verified=0