Update patch set 2
Patch Set 2: Code-Review+2

This is an acceptable way to do this, so I am +2 and merging.  However, I would like to open the conversation as to whether or not uint16, rather than a semantic approach (like SYN, ACK, RST, etc) might be easier for the user and clearer in general.  Its not a nit I think should be blocking, but I would encourage the submitter to think about it and possible discuss it :)

Patch-set: 2
Reviewer: Gerrit User 9 <9@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+2
Label: SUBM=+1