Update patch set 1
Patch Set 1:

@Ed - Of course, I likely should have linked to some resources earlier.

This very popular Stack Overflow question[1] has a succinct, clear explanation of some major reasons not to use which.

Regarding `which` spawning a new process:

[fedora@dfarrell-opendaylight-fedora ~]$ type -a which
which is aliased to `alias | /usr/bin/which --tty-only --read-alias --show-dot --show-tilde'
which is /usr/bin/which
[fedora@dfarrell-opendaylight-fedora ~]$ type -a command
command is a shell builtin
command is /usr/bin/command
[fedora@dfarrell-opendaylight-fedora ~]$ echo $SHELL
/bin/bash

So on this system (Fedora 20), given the bash shebang at the top of the run.sh script, which will resolve to /usr/bin/which (spawns a new process) and `command` will use a shell builtin.

I don't know of a good resource to link to with regard to which packages ship with various distros, but if you can spin up a Fedora 20 Cloud[2] image and check for which, you'll see that it doesn't ship with it. If you can't easily do that and don't want to take my word for it, I can spin up a fresh one on our OpenStack deployment and share the login info with you out-of-band.

[1]: http://stackoverflow.com/questions/592620/how-to-check-if-a-program-exists-from-a-bash-script
[2]: http://fedoraproject.org/en/get-fedora#clouds

Patch-set: 1
Reviewer: Gerrit User 1723 <1723@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0