Update patch set 12
Patch Set 12:

(1 comment)

Code changes in sal-remoterpc-connector sould be separete commit.

This patch also introduces new POM files, which were removed previously, so it seems it is not based on the master,
please could you actualize it based on current state of the master.

Also this patch will introduce unified version for all components which actually breaks semantic versioning.

Patch-set: 12