Update patch set 5
Patch Set 5:

Mathieu,

actually on this topic, i did quite a lot of investigation when doing the scoping on what to pick as web infra. And the OSGi way to do web services is via Web Application Bundle:

http://www.osgi.org/download/r5/osgi.enterprise-5.0.0.pdf

page 449.

But quite frankly there is not much difference in technique between the pax-web portion and the virgo way. The reason for choosing tomcat, was practical. The presence of valve as cental way of enforcing policies, this because i didn't know about servlet 3.0 goodies that Prasanth mentioned. Now i would suggest, we merge this one and we start a work on redesigning this part to allow tomcat and jetty and any other replacement if needed. I love the fact that servlet contract has different implementations so we can pick and choose the more proper for the environment we run into.

My 2 cents,
Giovanni

Patch-set: 5
Reviewer: Gerrit User 19 <19@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0