Topology manager: get rid of sychronized sections
This path changes the way MD topology manager works by introducing an
internal queue, which is filled from notifications. The queue is emptied
by a background thread which manages bulk&batch operations. This reduces
pressure on the datastore, which has less transactions to deal with, and
also provides better handling of notification -- they do not block, but
run concurrently as long as the queue is not full.

Change-Id: I902c9a4d6e2c4b85eb9ae8eaa5a543b6e602ad48
Signed-off-by: Robert Varga <rovarga@cisco.com>