DO NOT MERGE - Clustered Datastore POC using Akka continuation- Fault-tolerance of Persisted Actors
This POC code is to validate the application of 1-for-1 strategy of Akka fault-tolernace wherein
persisted state of an actor representing one Shard is not impacted, when a crash of another shard occurs and it is restarted by ShardManager.
Moreover it demonstrates having a MonitorActor to which ShardManager delegates the responsibility of monitoring  Shard Actors being up  and notifying it
when any Shard Actor is down.

We have two actors (TreeShard, FamilyShard) representing two shards in the system. When the program is run using the scripts/run-supervisor-poc.sh, both the shards are populated with some data and then periodically FamilyShard goes down and gets restarted by ShardManager. On restart FamilyShard state gets constructed properly from the Journal.The TreeShard state is also displayed to show it is not impacted.

Usage:
Go to md-sal/poc/akka/scripts folder and run ./run-supervision-poc.sh - this will build the .jar and run the command to demo the POC.

Change-Id: Ibe8a057f45afecd8c49fe453f3a0260db6157bed
Signed-off-by: Basheeruddin Ahmed <syedbahm@cisco.com>