Update patch set 1
Patch Set 1:

I really like all of Devin's comments, but I want to add a few high-level questions I have that I still can't answer based on this.

First, what are the exact properties the transactions are providing? The worry I have is that most people think ACID when they think transaction and I don't think we can provide that.

For ReadOnlyTransactions, it's pretty clear to me what is meant. This is a snapshot read. It might not get the latest version of the data, but it will return reads such that they are all from a certain instant in time. That is, there is a complete ordering of transactions including this one such that all reads from this transaction return the effect all writes from transactions that "occurred" before it and none of the effects of writes from transactions that "occurred" after it.

Note that these semantics do not require that different ReadOnly transactions go forward in time. It's entirely possible for me to use one ReadOnly transaction, finish with it and then get another and have the second give reads for an older point in time than the first. To avoid this we would need to do one of:
(i) to change the semantics to always return the latest values, which effectively means locking out writes,
(ii) provide a notion of logical time and allow reads to specify that they don't want to read any values older than a given value as well as providing a way to figure out what the oldest read from a transaction was when you close it, or
(iii) force the newer transactions never return older information than older ones in the factory.

The details we'd need from semantics of what we can expect from write-only semantics are similar to the above and for read-write, they are even more complex.

This is harder still when we consider that we might have the data store span multiple devices, some of which might not even be the controller itself, e.g., when mounting devices via NETCONF. Is the intention to have transactions span the different distributed parts of the data store? And provide the same semantics?

Patch-set: 1
Label: Verified=0