Update patch set 6
Patch Set 6: Code-Review-1

(53 comments)

At a high-level this is a huge improvement and the examples really help drive points home.

Comments fall into 3 categories:
* Add <p> tags to get paragraphs to format right.
* Really minor nits which start with "Really minor nit". If you want to change them great, otherwise I'm happy to fix them later myself in a separate patch.
* Just normal comments.

At a high level, I think we could still use to explain the semantics of put vs. merge, but I don't want to block this patch on in.

Also, I'm unclear why a write-only transaction should ever fail unless, maybe, if because of a change in the tree above the changes will not be visible as these nodes are no longer part of the data tree. In all other cases, there's no way for the code creating the write-only transaction to know what any of the values it's writing were before so, it shouldn't care if they changed in the interim. Am I missing something?

Patch-set: 6
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1