Update patch set 4
Patch Set 4:

this might not be the right place to start this conversation, but I've been thinking of how to represent shards and I'm curious if it doesn't make sense to actually represent them as part of the YANG model itself.

We could have a universal augmentation called "shard" which could then be attached to any node in the data tree to represent that the subtree starting there should be a shard. Obviously, the shard would end whenever you hit a data tree node augmented with "shard" or you get to a leaf.

Patch-set: 4
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0