Added an AbstractLazyNotificationTransformer to sal-binding-utils
Its a pretty common use case for a provider to consume
some notifications and emit others.  Its also pretty common
that it doesn't need to consume and process notifications
unless another module is subscribed to the notifications
it emits.  This AbstractLazyNotificationTransformer codifies
that logic in a manner that allows for folks to do a simple
extention of it to get the job done, without worrying about
the underlying logic of the Laziness.

Integration tests are included in sal-binding-it.

Change-Id: I09ef50e805b30d098f0239bf913daebdc8422b03
Signed-off-by: Ed Warnicke <eaw@cisco.com>