Update patch set 5
Patch Set 5:

I have some general questions here - 

Is a transaction really intended to be used by more than one thread? It seems logical to me that require that someone using a transaction does so from a single thread. If so, we should document this in the class (otherwise the put, delete, and merge likely need to be synchronized well, all depending on their underlying implementation).

Also, I am confused about the "cancel" method here (which seems to be the only reason why we need to synchronize in the first place). I would like to suggest the following changes:

1) remove the cancel method, and all synchronization / volatile
2) To cancel a transaction that you started modification on, but don't want to commit, call "close" without calling commit.
3) To cancel a transaction which you are in the process of committing, call cancel on the returned Future.

If we make these changes, along with documenting the fact that this call needs to be used from a single thread, then I think we can reduce the complexity of the class and remove all concurrency concerns.

Patch-set: 5