Update patch set 3
Patch Set 3:

Re the "// FIXME: This call should not be blocking.".

I think Tom's point (and I agree) is that we should add a little more information when putting in TODO's like this. Even if we plan to fix this tomorrow, adding some more information here would help folks who are reading this code to understand what needs to change. Even something as simple as "FIXME: This call should not be blocking. We should be using the ListanbleFuture.addListener method to offload this call to another thread" etc is short, but descriptive and explains what needs to happen. That way someone else could make the change if you don't have time to get to it.

Patch-set: 3