Update patch set 2
Patch Set 2:

So... the reason this was done with integration vs unit tests is that the interesting thing here is:

Do we do the right thing subscription-wise for notifications.  For some of these contracts, I probably *could* do it with unit tests... but for some (like the case of registering a LazyNotificationTransformer *after* someone has subscribed to it, you really *do* need an integration test to check that, as it depends on the MD-SAL behavior.

Integration tests *are* run by Jenkins at verify and merge... so they *are* checked.

Patch-set: 2