Bug-548:Remodelling of the queue constructs.
Currently the queue models are not properly designed. For e.g. The properties for a queue need to be a list. Also the queue config request has to be evaluated as configuring queues from controller is not supported by Openflow.
There are lot of duplicate models as well (e.g queue-management). This patch remodel the queues. Regression is minimal as statistics is the only component which gets the queue-stats and uses these models.The new models would serve as the base for implementing queue-config fetch request and to be stored in the operational datastore (whenever its implemented).

Change-Id: I65b71f4e8f36626237c484fb50bd2b1f19459ba4
Signed-off-by: Kamal Rameshan <kramesha@cisco.com>