Update patch set 3
Patch Set 3:

Patch Set 3 contains fixes for two comments. On going question is documenting the node-connector - whether we think the node id belongs in the node connector id. 

Here is my latest response:

What we do today for openflow is myprotocol:<node-id>:<connector-id> but in talking with Ed we felt this was unnecessarily redundant because you always access a node connector in the context of a node already. for example:
openflow:12/openflow:12:3
In this case the thought was that the second ":12" was not necessary.

Patch-set: 3
Reviewer: Gerrit User 1840 <1840@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0