Update patch set 1
Patch Set 1:

(4 comments)

Few comments inline. Also, anywhere where we have have @Deprecated, we should provide a quick description as to why, and the recommended method to use (right now you have the recommended implementation, but doesn't help show the user the "right" way to do it, or why that way is no longer correct).

Patch-set: 1