Update patch set 1
Patch Set 1:

So I understand why we would want whitespace included automatically from a developers POV (I can just query a URL and get back pretty XML/JSON). However I would actually argue that default should be no white space because that is more performant (since we are sending less text back with formatting). Additionally this is a parameter that is not part of the restconf draft, which is fine, but means folks will not necessarily know it exists and would definitely not know that they have to pass it everytime to get max performance benefit.

So I would like to propose that we swap the defaults and have it default to true (i.e. output should NOT be formatted) unless you specifically add the ignoreWhiteSpace argument.

Patch-set: 1
Label: Verified=0