Update patch set 3
Patch Set 3:

These renames look good. My only question is can we drop the word "Data" from the ReadTransaction, ReadWriteTransaction and Write Transactions? And simplify to just the type of transaction? I understand you are trying to maintain a tie between the transaction type and the other Binding Aware interfaces, but I would be a fan of simplifying those ones even more, especially because they are always going to be created in the context of a DataBroker today.

Patch-set: 3
Label: Verified=0