Update patch set 6
Patch Set 6:

Tony - if possible please just add a quick additional comment in the close method with the FIX ME that you added there. That said, I don't think we should hold this demo up anymore for that one fix - so if Tony can push that before we merge the now dependent (parent) changes great - if not we can push this when the parent change to the DataBroker names can get pushed.

Patch-set: 6
Label: Verified=0