Update patch set 2
Patch Set 2:

It looks like the existing default behavior (prior to this change) was to pretty print the response XML / JSON respectively. While I agree that the default should be to not pretty print we probably do want to send mail to the community if we want to change that default behavior with this bug.

Looking back at the bug (1016) it was really about just adding a newline to the end of generated output. The addition of a parameter to turn it off is a side benefit (an enhancement beyond what we get today). So I think a default= false is OK (even though I would prefer it to be true :) ). We can open a second bug to change the default pretty-print behavior.

Patch-set: 2
Label: Verified=0