Update patch set 1
Patch Set 1: Code-Review-1

(8 comments)

Pretty clean. We will definetly interfere with the renaming patch, so one of them will need to be rebased depending on which gets pushed first.

A few comments inline ( mostly asking for links to the SPI contact, or at least more info). Also, when deprecating the transactions, it may be worth providing a link to the data broker interfaces where you create those transactions, so people get the idea about the new way to build those as well. This second part is not as critical, would just be a nice helper.

Patch-set: 1
Label: Code-Review=-1