Update patch set 3
Patch Set 3: Code-Review-1

While the changes do handle the case where multiple child container nodes are specified, as in the unit test, it doesn't address the case for which the bug was created, ie where a yang container is specified as a json array:

      "ero":
      [ 
        { 
          "subobject":
          { 
             ...
          }
        },
        { 
          "subobject":
          { 
             ...
          }
        }
      ]

In this case, the JsonReader sees the "ero" element as a JSON array and when it gets the element as a JSON array, the json parser only returns 1 of the "subobject" elements. This is the reason the bug creator only got 1 hop created instead of 3. The parser allows this syntax although I'm not sure if it's correct. 

The correct syntax corresponding to the schema is:

      "ero":
      { 
        "subobject":
         [
           { 
              ...
           },
           { 
             ...
           }
         ]
      }
 
This is tricky. It seems the validation to detect the former as a bad element needs to be done in the JsonReader, as it's specific to json syntax, but it would need to have knowledge of the schema. 
 
I reproduced this by adding the following to the toaster's make-toast input:

        container ero {
           list sub {
              leaf lf { type string; }
           }
        }

then sending paylod:

{ 
  "input" :
  {  
     "toaster:toasterDoneness" : "3",
     "toaster:toasterToastType":"wheat-bread",
     
     "ero":
     [ 
       { 
         "sub" : {
           "lf" : "1"
         },
         "sub" : {
           "lf" : "2"
         }
       }
    ]
  }
} 

The RPC will succeed but in the output from the toaster you will only see 1 of the "sub" elements.

Patch-set: 3
Label: Code-Review=-1