Update patch set 3
Patch Set 3: Code-Review-1

(9 comments)

The current implementation looks great and functions as expected. One minor comment inline to make the parsing of the param a little easier / more resilient, but that isn't blocking.

The -1 comes from the fact that we are changing the default behavior, which is fine, however there are a number of other methods that return structured data objects which would now loose the ability to get pretty-printed data. I can see at least 2 ways forward:

1) Change the other methods in the RestconfImpl.java to parse and return pretty printed data in their structured Data classes
2) Change the default back to pretty print by default (since that is what was in there before) so we don't loose any functionality. Then create a new defect to change the default behavior for everything. The original defect was really just about adding a newline to the end of hte pretty printing - everything else has been additional add-ons (which is awesome). :)

My preference is #1 - but that may be a bit more work than you bargained for. If we do do #1, we will want to send a warning e-mail to the community to let them know that we are changing default behavior (as a courtesy).

Thanks

Patch-set: 3
Label: Code-Review=-1