Update patch set 8
Patch Set 8: Code-Review+1

Looks good! I like the short renames.

I am concerned about some interfaces that were generated in the toaster example that have the same name ("DataBroker"). I am not sure if that is because of the way the toaster ties into the config system, or some other reason.

Either way, that does not block this. Can push once someone else +1 (I think it is worth having a few opinions on this).

Patch-set: 8