Bug 1027: Sort map of instance-identifiers and data by length.
In order to have more reliable deserialization of Maps
of Bidning-Independented data to Binding-Aware
entries should be sorted by their length, with shorter
ones beforehand, since codecs for them have higher
probability to be known.

Parent codecs also contains references to child-classes
which may be never used in new context so this
makes deserialization more predicable and easier to debug.

Change-Id: Iad1490f893e946a2a019c88570325dcfb11c244d
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>