Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

Was about to submit and I realized that a lot of the links point to BindingDataBroker, but we have renamed that to DataBroker. So we need to update those links. I pointed out one specific case as an example.

Also two additional minor things if we can fix them quickly would be great.

Patch-set: 3
Label: Code-Review=-1