Update patch set 1
Patch Set 1:

It looks like you tried to fix this with commit:
https://git.opendaylight.org/gerrit/#/c/8394/

It can't be merged because it depends on this commit.  The correct way to fix a broken commit is to 

commit --amend

the broken commit, not create a new one.

Hope that helps :)

Patch-set: 1
Reviewer: Gerrit User 9 <9@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0