Introducing simple merge strategy for config subsystem
MergeEditStrategy wasn't really merging.
Now it is, but only for ObjectName[].
However, since ObjectName[] is 90% of the interesting
cases, this should get us a long way.

This is needed to be able to handle reasonably
extension, with the ability to wire in multiple
extension modules independently, as we need to do
for things like OF extension.

Change-Id: Ieca579cbc6781d77f489f5f754490c8da8f65188
Signed-off-by: Ed Warnicke <eaw@cisco.com>