Update patch set 1
Patch Set 1:

So... personal aesthetic nit, as a personal matter, I would tend to prefer an additional signature of:

void put(LogicalDatastoreType store, InstanceIdentifier<?> path, DataObject data, boolean createMissingParents);

(in addition to the existing put method)

(it closer to the mkdir -p example)

The question then becomes what the default behavior for the simple put is... I would tend to advocate for the default behavior being createMissingParents = true.  I feel this because it is likely to be the most well understood concept for most folks, and they are likely to run into circumstances otherwise where they don't quite get the expected behavior.

Thoughts from others?

Patch-set: 1