Update patch set 1
Patch Set 1:

I like the idea of adding a boolean parameter. 

How about not adding separate put/merge methods? There would only be the methods with the boolean param. That takes the question of default behavior essentially out of the picture and forces the user to think about it and choose.

On a side note, why is the 'create parents' option only on the ReadWriteTransaction and not on WriteTransaction?

Patch-set: 1
Label: Code-Review=0