Update patch set 1
Patch Set 1:

I think we should be consistent with the way it did work before the change. I think we should create the parents by default, and provide an option to optionally fail if the parents don't exist.

I think 80% of the time people would want to use it in this way (makes it easier to use).

Patch-set: 1