Update patch set 1
Patch Set 1:

I second Tom's thoughts here. We really want one method, which should be the type safe method. Having the two will likely confuse most people.

For the same reason that we renamed BidningDataBroker to DataBroker, and we are "hiding" the BIDataBrokers.. we should try and take the same approach here.

Is there anyway to just have the type safe version?

Patch-set: 1
Label: Verified=0