Update patch set 1
Patch Set 1:

I think the only way is to remove the type-unsafe method that is defined in the base AsyncReadTransaction interface. I understand having a generalized base interface that is common between binding and BI and, if the generics work effectively to define common coherent methods, then that's fine. But in this case the generalized base method is not ideal and we shouldn't let that yield a binding interface that isn't coherent.

Patch-set: 1