Update patch set 1
Patch Set 1:

Robert - that was my initial thought too, remove the base interfaces and generics to make it simpler. The generalized base classes can be pushed to the impls. The only reason I can see for keeping the base interfaces  is that they provide a place to put common javadocs between BA and BI so we don't have to duplicate them. Of course we could put the complete javadocs in the BA interfaces and link from the BI interfaces since BI will be less commonly used.

Patch-set: 1