Update patch set 1
Patch Set 1:

Based on all of the errors that are out there and questions there was clearly a given behavior that pre-existed prior to this change. I think we should revert the behavior to what existed before since that is the expectation that everyone has thus far.

My understanding of the existing behavior is:

1) If parent nodes don't exist, create them.
2) Users must explicitly delete the parent nodes.

Is that how it worked before? Regardless, we should just make this work like it did before so all code that made those assumptions start working again.

Patch-set: 1
Reviewer: Gerrit User 1840 <1840@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0