Update patch set 1
Patch Set 1: Code-Review-1

(14 comments)

In addition to changes you made, I'd like to also change the error message in ControllerContext#collectPathArguments line 699 to indicate the key is missing in the URI:

  "Missing key value for list element '" + listNode.getQName().getLocalName() + '" in the URI."

Also the error tag should be DATA_MISSING.

In the future, please avoid reformatting an entire file. This makes it harder to review and know what the actual changes were. In RestconfImpl there were a lot of extraneous formatting changes not related to the defect changes. If you're using Eclipse with the Format on save action, please turn that off.

Patch-set: 1
Label: Code-Review=-1