Update patch set 1
Patch Set 1:

Tony - if we're going to remove common base interfaces then I think we should do it for all to be consistent, including WriteTransaction.

Re: javadocs, the BA DataBroker API will be the commonly used one so I'd suggest to put the rich docs in those interfaces and put minimal in BI interfaces with links to BA interfaces to avoid duplication.

Patch-set: 1
Label: Code-Review=0