Bug 1392: Change ReadTransaction#read to return CheckedFuture
Added a ReadFailedException that is used with the CheckedFuture.

Moved the RpcError list from TransactionCommitFailedException to a base
OperationFailedException and also derived ReadFailedException from
OperationFailedException.

Added a ReadFailedExceptionMapper class and created a generalized base
ExceptionMapper class. Also derived
TransactionCommitFailedExceptionMapper from ExceptionMapper..

In addition, Futures#addCallback and Futures#transform call Future#get but
CheckedFuture doesn't override the get methods to translate exceptions
to the checked type. This kind of defeats the purpose of CheckedFuture.

So I added a MappingCheckedFuture class that overrides the get methods
to translate exceptions. Instead of using Futures#makeChecked in the
read, I create a MappingCheckedFuture. I also changed the write Tx
submit to use MappingCheckedFuture.

Change-Id: I5c4f717f0b8664b7d39c1e6f0366525f04e6634d
Signed-off-by: tpantelis <tpanteli@brocade.com>