Update patch set 3
Patch Set 3:

(1 comment)

Using Strings as identifiers would be just fine - it's opaque enough. I don't look for any special meaning in the identifier. When it is used in logging it will make it easier for someone to follow the logs.

It's normal for messages to not extend a common message type. Though in later commits I did add a marker type for all the Raft rpc's.

Patch-set: 3
Label: Verified=0