BUG 932 - Swagger HTTP POST contains incorrect object
POST link now contains specific JSON structure which ends with POST suffix
and contains all containers and lists which are direct subchildren of
current node (URI link). It is list of possible objects which can be used
in POST request. Concretelly are these objects specified below this *POST
JSON object and are also all specified in Parameters section of GUI POST
URI link with ** prefix. Two stars (**) mean that only one of parameters
with ** prefix should be specified.

Change-Id: I3acdfd1f181fbef7cca0060d534b706d934bb1e3
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>