Update patch set 1
Patch Set 1: Code-Review-1

Tony - I am not sure how this validates the bug since it appears to only write to the data store.

Can you please add comments to the test to explain how this is validating bug 1352?

Also, I am not a fan of having the bug number in the name of hte test. I would rather the class name describe the class or functionality under test and the bug reference being comments or test name....

Patch-set: 1
Label: Code-Review=-1