Update patch set 1
Patch Set 1:

Added Ramkumar as we reviewer - I want to make sure we covered the scenarios he describes in Bug 1333 adequately.

Also, this is a regression test - it is unclear if no change was necessary and this test is suppose to prove it? Or was another change to fix this added somewhere else? If so please provide a pointer to it.

Patch-set: 1